Revert "initial FluidSynth support" (for now at least, to get my fork working and fix MIDI input)
This reverts commit 8b7fe147384073bcf47e13676b95aab4445484be.
